Add support for triggering side effects by returning the list of operations applied after integration.
When you receive a remote operation and integrate it,
the integrate method now returns a vector of operations and the updated WString.
Mostly you expect the vector to contain just the operation you integrated.
However, it could be that the operation could not be integrated, and was added to a queue.
Or it could be the operation unblocked the queue, allowing more than one operation to apply.
By returning the vector of operations a client can run side-effects,
such as putting a character on the screen.